NOJIRA-Fix-call-bridge-auto-answer: move ARIChannelStateChange to existing ari_event.go
- bin-call-manager: Merge ARIChannelStateChange into confbridgehandler/ari_event.go (existing file)
- bin-call-manager: Remove incorrectly created separate arievent.go file
